Formulario de nuevo kanji

+++ src/app/shared/components/tag/tag.component.ts
@@ -1,17 +1,21 @@
 import { CommonModule } from '@angular/common';
 import { Component, Input, inject } from '@angular/core';
+import { Generic } from '@models/generic.model';
 import { TagConfiguration } from '@models/tagConfiguration.model';
 import { GenericService } from '@services/nihongo/kanji/generic.service';
+import { ToastrService } from 'ngx-toastr';
 import { firstValueFrom } from 'rxjs';
 
 @Component({
   selector: 'tag',
   standalone: true,
   imports: [CommonModule],
-  templateUrl: './tag.component.html',
-  styles: ``,
+  templateUrl: './tag.component.html'
 })
 export class TagComponent {
+
+  private _toast = inject(ToastrService);
+
   @Input() configuration!: TagConfiguration;
   showInput = false;
 
@@ -19,31 +23,46 @@ export class TagComponent {
 
   async save(value: string) {
     try {
-      this.showInput = false;
 
-      this.configuration.actual = {
-        kanjiId: this.configuration.parentId,
-        value,
-        name: value,
-      };
+      let exist = this.configuration.data?.some(item => item.value == value);
 
-      this.configuration.data?.push(
-        await firstValueFrom(this._genericService.post(this.configuration))
-      );
+      if (!exist) {
+        this.showInput = false;
+
+        if (this.configuration.parentId) {
+
+          this.configuration.actual = {
+            kanjiId: this.configuration.parentId,
+            value,
+            name: value,
+          };
+
+          this.configuration.data?.push(
+            await firstValueFrom(this._genericService.post(this.configuration))
+          );
+        } else {
+          this.configuration.data?.push({ value, name: value });
+        }
+      } else {
+        // Agregar mensaje de elemento repetido
+        this._toast.warning(`"${value}" ya se encuentra en la lista`, 'Elemento repetido');
+      }
     } catch (e) {
       console.error(e);
     }
   }
 
-  async delete(id: number | undefined) {
+  async delete(tag: Generic) {
     try {
-      this.configuration.actual = {
-        id,
-      };
+      if (this.configuration.parentId) {
+        this.configuration.actual = tag;
+        await firstValueFrom(this._genericService.delete(this.configuration));
+        this.configuration.data = this.configuration.data?.filter(item => item.id != tag.id);
+      } else {
+        this.configuration.data = this.configuration.data?.filter(item => item.value != tag.value);
+      }
 
-      await firstValueFrom(this._genericService.delete(this.configuration));
-
-      this.configuration.data = this.configuration.data?.filter(item => item.id != id);
+      this._toast.success('Elemento eliminado');
 
     } catch (e) {
       console.error(e);
